refactor(oscript): собрать oscript-пакеты под types.oscript
types.inferencer.annotations → types.oscript.annotations,
types.inferencer.autumn → types.oscript.autumn: всё oscript-зависимое — в
пакете oscript, библиотеко-специфичное — в его подпакетах (annotations,
autumn, extends_); зависимости между пакетами прослеживаются по импортам.
У каждого пакета package-info.java с javadoc и @NullMarked.
