+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/types/oscript/autumn/AutumnAnnotations.java
@@ -19,9 +19,9 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with BSL Language Server.
  */
-package com.github._1c_syntax.bsl.languageserver.types.inferencer.autumn;
+package com.github._1c_syntax.bsl.languageserver.types.oscript.autumn;
 
-import com.github._1c_syntax.bsl.languageserver.types.inferencer.annotations.OScriptAnnotations;
+import com.github._1c_syntax.bsl.languageserver.types.oscript.annotations.OScriptAnnotations;
 import lombok.experimental.UtilityClass;
 
 /**
