+++ src/test/java/com/github/_1c_syntax/bsl/languageserver/codelenses/BeanUsagesCodeLensSupplierTest.java
@@ -27,9 +27,9 @@ import com.github._1c_syntax.bsl.languageserver.context.DocumentContext;
 import com.github._1c_syntax.bsl.languageserver.context.symbol.ConstructorSymbol;
 import com.github._1c_syntax.bsl.languageserver.context.symbol.MethodSymbol;
 import com.github._1c_syntax.bsl.languageserver.context.symbol.SymbolTree;
-import com.github._1c_syntax.bsl.languageserver.types.inferencer.autumn.AutumnBeanIndex;
-import com.github._1c_syntax.bsl.languageserver.types.inferencer.autumn.AutumnBeanIndex.FactoryMethodBean;
-import com.github._1c_syntax.bsl.languageserver.types.inferencer.autumn.AutumnNavigation;
+import com.github._1c_syntax.bsl.languageserver.types.oscript.autumn.AutumnBeanIndex;
+import com.github._1c_syntax.bsl.languageserver.types.oscript.autumn.AutumnBeanIndex.FactoryMethodBean;
+import com.github._1c_syntax.bsl.languageserver.types.oscript.autumn.AutumnNavigation;
 import com.github._1c_syntax.bsl.languageserver.utils.Resources;
 import com.github._1c_syntax.utils.Absolute;
 import org.eclipse.lsp4j.CodeLens;
